% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPSFS-package.R
\docType{package}
\name{PPSFS-package}
\alias{PPSFS}
\alias{PPSFS-package}
\title{PPSFS: Partial Profile Score Feature Selection in High-Dimensional Generalized Linear Interaction Models}
\description{
This is an implementation of the partial profile score feature selection (PPSFS) approach to generalized linear (interaction) models. The PPSFS is highly scalable even for ultra-high-dimensional feature space. See the paper by Xu, Luo and Chen (2022, \doi{10.4310/21-SII706}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/paradoxical-rhapsody/PPSFS}
  \item Report bugs at \url{https://github.com/paradoxical-rhapsody/PPSFS/issues}
}

}
\author{
\strong{Maintainer}: Zengchao Xu \email{zengc.xu@aliyun.com}

Authors:
\itemize{
  \item Shan Luo
  \item Zehua Chen
}

}
\keyword{internal}
