% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppsfs.fit.R
\name{ppsfs.fit}
\alias{ppsfs.fit}
\title{Partial Profile Score Feature Selection for GLMs}
\usage{
ppsfs.fit(
  x,
  y,
  family,
  fitFun = glm.fit,
  ...,
  keep = NULL,
  maxK = NULL,
  verbose = FALSE
)
}
\arguments{
\item{x}{Matrix.}

\item{y}{Vector.}

\item{family}{See \link{glm} and \link{family}.}

\item{fitFun}{Fitting function, by default \link{glm.fit}.}

\item{...}{Additional parameters for \link{glm.fit}.}

\item{keep}{Initial index set of features that are included in model fitting.}

\item{maxK}{Maximum number of identified features.}

\item{verbose}{Print the feature path?}
}
\description{
This is a simplified but more stable implementation for main-effects selection
compared to \link{ppsfs}.
It removes the standaridization, but adds the weights in the partial profile score.
}
\details{
That \code{ppsfs(x, y, family="gaussian")} is an implementation to
sequential lasso method proposed by Luo and Chen(2014, \doi{10/f6kfr6}).
}
\examples{
set.seed(2025)
n <- 300
p <- 1000
x <- matrix(rnorm(n*p), n)
eta <- drop( x[, 1:3] \%*\% runif(3, 1.0, 1.5) )
y <- eta + rnorm(n, sd=sd(eta)/5)
print( A <- ppsfs.fit(x, y, 'gaussian', verbose=TRUE) )

}
\references{
Z. Xu, S. Luo and Z. Chen (2022). Partial profile score feature selection in
high-dimensional generalized linear interaction models.
Statistics and Its Interface. \doi{10.4310/21-SII706}
}
